+++ app/Config/Routes.php
@@ -78,6 +78,7 @@ $routes->group('operativo', function ($routes) {
         $routes->get('(:num)',         'Operativo\InterventiController::show/$1');
         $routes->get('(:num)/edit',    'Operativo\InterventiController::edit/$1');
         $routes->post('(:num)/update', 'Operativo\InterventiController::update/$1');
+        $routes->post('(:num)/chiudi', 'Operativo\InterventiController::chiudi/$1');
         $routes->post('(:num)/delete', 'Operativo\InterventiController::delete/$1');
     });